+++ CHANGELOG.md
@@ -37,6 +37,15 @@ All notable changes to this project will be documented in this file. For commit 
  - Playback queue wasn't updating when changing of folder (#2360)
  - Added missing `UserDefaults.ShowCopyPath` option to config file (#2364)
 
+## v1.3.10
+
+ **Notes**:
+ - When indexing is disabled for a source, the usage will always reported as partition size.
+ - removed deprecated `source.config.disableIndexing`, see [rules](https://filebrowserquantum.com/en/docs/advanced/source-configuration/conditional-rules/#disable-indexing)
+
+ **BugFixes**:
+ - Disable index option not working (#2385)
+
 ## v1.3.9
 
  **Security**: